n/a

+++ 20210129/20210129_Task10.c
@@ -53,7 +53,7 @@ double LeastCommon(double n1, double n2){
     max = (n1 > n2) ? n1 : n2;
 
     while (1) {
-        if ((max%n1) == 0 && (max%n2) == 0) {
+        if (((int)max%(int)n1) == 0 && ((int)max%(int)n2) == 0) {
             printf("The LCM of %d and %d is %d.\n", n1, n2, max);
             break;
         }
